chore: Add Babel and ESLint configurations***
This commit adds the .babelrc and .eslintrc.json files with the necessary configurations for Babel and ESLint. These configurations include presets and plugins for Next.js and core-web-vitals. The root directory for the Next.js app is set to "apps/evo-editor/". Additionally, the package.json file is updated with devDependencies for eslint-config-next and typescript.